# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arcane Waste Management API Documentation",
        "description": "Media endpoints for media upload and management.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://arcane-2nd-version.test/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/company/general-settings": {
            "get": {
                "tags": [
                    "General Settings"
                ],
                "summary": "Get general settings",
                "description": "Fetch general settings for the authenticated company. Supports filtering, sorting, and optional pagination.",
                "operationId": "6a153c5fed6fe5e3df4de2274b6ce0be",
                "parameters": [
                    {
                        "name": "filters[key]",
                        "in": "query",
                        "description": "Filter settings by key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[value]",
                        "in": "query",
                        "description": "Filter settings by value",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[type]",
                        "in": "query",
                        "description": "Filter settings by type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order: 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Whether to paginate the results (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page if pagination is enabled",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number if pagination is enabled",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "General settings fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "General settings fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GeneralSettingResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "links": {
                                                    "properties": {
                                                        "first": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/general-settings?page=1"
                                                        },
                                                        "last": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/general-settings?page=5"
                                                        },
                                                        "prev": {
                                                            "type": "string",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "next": {
                                                            "type": "string",
                                                            "example": "http://example.com/api/general-settings?page=2",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/company/general-settings/{id}": {
            "get": {
                "tags": [
                    "General Settings"
                ],
                "summary": "Get a single general setting",
                "description": "Fetches a single general setting by its ID for a given model (polymorphic).",
                "operationId": "8aecb4424b8de6c04dbad289453229a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the general setting",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "General setting fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "General setting fetched successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GeneralSettingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "General setting not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "General setting not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/company/general-settings/{id}/{key}": {
            "put": {
                "tags": [
                    "General Settings"
                ],
                "summary": "Update a general setting",
                "description": "Updates a specific general setting by its ID and key for the authenticated company.",
                "operationId": "6fa171c0bcbb5e97012230d1fabffe60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the general setting",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Key of the general setting to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "order_cancelation_time_limit"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string",
                                        "example": "30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "General setting updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "General setting updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GeneralSettingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "General setting not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "General setting not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GeneralSettingResource": {
                "title": "General Setting",
                "description": "General setting resource response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "key": {
                        "type": "string",
                        "example": "key"
                    },
                    "value": {
                        "type": "string",
                        "example": "value"
                    },
                    "type": {
                        "type": "string",
                        "example": "type"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}